+++ src/bot/handlers/command_application.py
@@ -1,5 +1,5 @@
 from pydantic import ValidationError
-from telegram import BotCommandScopeChat, ReplyKeyboardRemove, Update
+from telegram import ReplyKeyboardRemove, Update
 from telegram.ext import (
     Application,
     CommandHandler,
@@ -9,27 +9,12 @@ from telegram.ext import (
     filters,
 )
 
-from bot.constants import button, text
-from bot.constants.state import CHECK, GET_USER_QUESTION
+from bot.constants.text import BACK_TO_MENU, STOP_MESSAGE
 from bot.keyboards.keyboards import main_menu_markup
 from bot.utils import QuestionModel, send_email
 
 
-async def greeting_callback(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> int:
-    '''
-    Базовая функция начинающая диалог с юзером
-    и открывающий доступ к check_secret_conv_handler.
-    '''
-    await context.bot.set_my_commands(
-        [button.START_CMD, button.HELP_CMD],
-        scope=BotCommandScopeChat(update.effective_chat.id),
-    )
-    await update.message.reply_text(text.START_MESSAGE_PART_ONE)
-    await update.message.reply_text(text.START_MESSAGE_PART_TWO)
-    return CHECK
+GET_USER_QUESTION = 1
 
 
 async def help_callback(
@@ -46,15 +31,6 @@ async def get_user_question_callback(
     '''Функция обрабатывает сообщение с вопросом от пользователя.'''
     user_id = update.effective_user.username
     user_question = update.message.text
-    if update.message.text.startswith('/'):
-        command = update.message.text[1:]
-        if command == 'menu':
-            await menu_callback(update, context)
-        elif command == 'stop':
-            await stop_callback(update, context)
-        elif command == 'start':
-            await greeting_callback(update, context)
-            return ConversationHandler.END
     try:
         QuestionModel(question=user_question)
     except ValidationError as e:
@@ -62,8 +38,8 @@ async def get_user_question_callback(
         error_message = error_message.replace('Value error,', '')
         await update.message.reply_text(error_message.strip())
         return GET_USER_QUESTION
-    subject = 'Обращение в поддержку телеграм бота'
-    body_text = f'Никнейм в телеграм: {user_id}\nВопрос: {user_question}'
+    subject = 'Вопрос от пользователя'
+    body_text = f'Пользователь: {user_id}\nЗадает вопрос: {user_question}'
     send_email(subject, body_text)
     await update.message.reply_text(
         'Ваш вопрос отправлен. Мы свяжемся с вами в ближайшее время.'
@@ -76,7 +52,7 @@ async def menu_callback(
 ) -> None:
     '''Команда перехода в главное меню.'''
     await update.message.reply_text(
-        text.BACK_TO_MENU, reply_markup=main_menu_markup
+        BACK_TO_MENU, reply_markup=main_menu_markup
     )
 
 
@@ -89,7 +65,7 @@ async def stop_callback(
     После её работы, бот будет принимать только команду /start.
     '''
     await update.message.reply_text(
-        text.STOP_MESSAGE, reply_markup=ReplyKeyboardRemove()
+        STOP_MESSAGE, reply_markup=ReplyKeyboardRemove()
     )
     return ConversationHandler.END
 
@@ -108,4 +84,3 @@ def register_handlers(app: Application) -> None:
     app.add_handler(help_handler)
     app.add_handler(CommandHandler('menu', menu_callback))
     app.add_handler(CommandHandler('stop', stop_callback))
-    app.add_handler(CommandHandler('start', greeting_callback))
